+++ navigator-compose/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Version `unknown`
+
+- **Fixed**: Proper lifecycle events on `Route`.
+
 ## Version `0.1-alpha30` _(2022-05-02)_
 
 - **Updated**: Jetpack Compose to version v1.1.1 (requires Kotlin v1.6.10).